Remove encoding/gob as a dependency
No longer use reflection as a way to send byte data.

A more performant way to encode the frame for RPC is to simply just
re-encode it as bytes using the encoder and decode it with the
appropriate library on the other end.

Some ideas:

For C#, one can use the machina library.
For JavaScript, one can write a go program using xivnet and compile
it as webasm.